#5 Adds .clang-tidy and revert tinyxml2 modifications

+++ external/tinyxml2/include/tinyxml2/tinyxml2.h
@@ -613,7 +613,7 @@ public:
     // p is the starting location,
     // the UTF-8 value of the entity will be placed in value, and length filled in.
     static const char* GetCharacterRef( const char* p, char* value, int* length );
-    static void ConvertUTF32ToUTF8( uint32_t input, char* output, int* length );
+    static void ConvertUTF32ToUTF8( unsigned long input, char* output, int* length );
 
     // converts primitive types to strings
     static void ToStr( int v, char* buffer, int bufferSize );
@@ -2378,4 +2378,3 @@ private:
 #endif
 
 #endif // TINYXML2_INCLUDED
-
